requirements.txt
asgiref==3.2.9
boto3==1.13.22
botocore==1.16.22
certifi==2020.4.5.1
chardet==3.0.4
coverage==5.1
dj-database-url==0.5.0
Django==1.11.29
django-forms-bootstrap==3.1.0
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
jmespath==0.10.0
Pillow==7.1.2
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
stripe==2.48.0
urllib3==1.25.9
